simple HEADLESS caveat
`try...except` on the required simulator import in `pixoo/__init__.py` to set  a `HEADLESS_NO_SIMULATOR` flag in the module to skip instantiating with the classes we're not importing. This is a workaround for skipping importing tkinter, like on a headless raspberry pi (no X windows and python-tk installed) 

I believe this closes the loop on the intended behavior according to the docs, stating `python-tk|python3-tk` is optional.